requirements.txt
streamlit
tensorflow
pillow
numpy
twilio
python-dotenv
